+++ tests/tt_metal/tt_metal/dispatch/dispatch_event/test_EnqueueWaitForEvent.cpp
@@ -12,7 +12,7 @@
 #include <tt-metalium/event.hpp>
 #include <tt-metalium/device.hpp>
 
-#include "hostdevcommon/cq_commands.hpp"
+#include "tt_metal/impl/dispatch/kernels/cq_commands.hpp"
 
 namespace tt::tt_metal {
