+++ tests/tt_metal/tt_metal/perf_microbenchmark/routing/test_mux_demux.cpp
@@ -6,7 +6,7 @@
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/rtoptions.hpp>
 #include <tt-metalium/device.hpp>
-#include "hostdevcommon/cq_commands.hpp"
+#include "tt_metal/impl/dispatch/kernels/cq_commands.hpp"
 #include "tt_metal/impl/dispatch/kernels/packet_queue_ctrl.hpp"
 #include "test_common.hpp"
 #include "routing_test_common.hpp"
